+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 <!-- There is always Unreleased section on the top. Subsections (Add, Changed, Fix, Removed) should be Add as needed. -->
 ## Unreleased
+- Add `validation {}` computation expression, accumulating failures of `let! ... and! ...` bindings of both `Validation` and `Result`
+- Add `Validation.zip`
 
 ## 2.0.0 - 2025-12-04
 - [**BC**] Use net10